Continuing cleaning up include setup.
