+++ benchmark/priority_queue.cpp
@@ -1,7 +1,7 @@
-#include <quicr/priority_queue.h>
+#include <quicr/containers/priority_queue.h>
+#include <quicr/quic_transport.h>
 
 #include <benchmark/benchmark.h>
-#include <quicr/quic_transport.h>
 
 static auto tick_service = std::make_shared<quicr::ThreadedTickService>();
