+++ benchmark/stream_buffer.cpp
@@ -1,7 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024 Cisco Systems
 // SPDX-License-Identifier: BSD-2-Clause
 
-#include <quicr/stream_buffer.h>
+#include <quicr/containers/stream_buffer.h>
 
 #include <benchmark/benchmark.h>
