+++ include/quicr/subscribe_track_handler.h
@@ -3,10 +3,10 @@
 
 #pragma once
 
-#include "base_track_handler.h"
 #include "containers/stream_buffer.h"
 #include "messages.h"
 #include "metrics.h"
+#include "track_handler.h"
 
 namespace quicr {
 
@@ -18,7 +18,7 @@ namespace quicr {
      *
      *  This extends the base track handler to add subscribe handling
      */
-    class SubscribeTrackHandler : public BaseTrackHandler
+    class SubscribeTrackHandler : public TrackHandler
     {
       public:
         /**
@@ -74,7 +74,7 @@ namespace quicr {
                               messages::FilterType filter_type,
                               const std::optional<JoiningFetch>& joining_fetch = std::nullopt,
                               bool publisher_initiated = false)
-          : BaseTrackHandler(full_track_name)
+          : TrackHandler(full_track_name)
           , priority_(priority)
           , group_order_(group_order)
           , filter_type_(filter_type)
