+++ src/CMakeLists.txt
@@ -8,7 +8,7 @@ else()
 endif()
 
 target_sources (quicr PRIVATE
-    base_track_handler.cpp
+    track_handler.cpp
     client.cpp
     ctrl_message_types.cpp
     ${ctrl_messages_SOURCE}
