+++ src/joining_fetch_handler.cpp
@@ -5,7 +5,18 @@
 
 #include <spdlog/spdlog.h>
 
+#include <memory>
+
 namespace quicr {
+    JoiningFetchHandler::JoiningFetchHandler(std::shared_ptr<SubscribeTrackHandler> joining_subscribe_handler)
+      : SubscribeTrackHandler(joining_subscribe_handler->GetFullTrackName(),
+                              joining_subscribe_handler->GetPriority(),
+                              joining_subscribe_handler->GetGroupOrder(),
+                              joining_subscribe_handler->GetFilterType())
+      , joining_subscribe_handler_(std::move(joining_subscribe_handler))
+    {
+    }
+
     void JoiningFetchHandler::StreamDataRecv(bool is_start,
                                              [[maybe_unused]] uint64_t stream_id,
                                              std::shared_ptr<const std::vector<uint8_t>> data)
@@ -41,17 +52,17 @@ namespace quicr {
                          obj.object_id,
                          obj.payload.size());
             try {
-                joining_subscribe_->ObjectReceived({ obj.group_id,
-                                                     obj.object_id,
-                                                     obj.subgroup_id,
-                                                     obj.payload.size(),
-                                                     obj.object_status,
-                                                     obj.publisher_priority,
-                                                     std::nullopt,
-                                                     TrackMode::kStream,
-                                                     obj.extensions,
-                                                     obj.immutable_extensions },
-                                                   obj.payload);
+                joining_subscribe_handler_->ObjectReceived({ obj.group_id,
+                                                             obj.object_id,
+                                                             obj.subgroup_id,
+                                                             obj.payload.size(),
+                                                             obj.object_status,
+                                                             obj.publisher_priority,
+                                                             std::nullopt,
+                                                             TrackMode::kStream,
+                                                             obj.extensions,
+                                                             obj.immutable_extensions },
+                                                           obj.payload);
             } catch (const std::exception& e) {
                 SPDLOG_ERROR("Caught exception trying to receive Joining Fetch object. (error={})", e.what());
             }
